+++ changelog.md
@@ -9,4 +9,5 @@
 * `.plugin` calls plugin function with `(app, options)` [semver-major]
 * Plugin function return value conform to object [semver-major]
 * Move utils to `lib/utils`
+* Tests refactor
 * Fix changelog